(#49) Move GitHub credential loading to github.go
