+++ main.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 	"gopkg.in/go-ini/ini.v1"
 	"os"
-	"os/user"
 	"path"
 	"path/filepath"
 	"regexp"
@@ -270,44 +269,6 @@ func locateProject(directory string) (string, error) {
 
 	// FIXME(#148): snitch is looking only for .snitch.yaml ignoring .snitch.yml
 	return path.Join(filepath.Dir(dotGit), ".snitch.yaml"), nil
-}
-
-func getGithubCredentials() (GithubCredentials, error) {
-	tokenEnvar := os.Getenv("GITHUB_PERSONAL_TOKEN")
-	xdgEnvar := os.Getenv("XDG_CONFIG_HOME")
-	usr, err := user.Current()
-
-	if err != nil {
-		fmt.Fprintln(os.Stderr, err)
-		os.Exit(1)
-	}
-
-	if len(tokenEnvar) != 0 {
-		return GithubCredentialsFromToken(tokenEnvar), nil
-	}
-
-	// custom XDG_CONFIG_HOME
-	if len(xdgEnvar) != 0 {
-		filePath := path.Join(xdgEnvar, "snitch/github.ini")
-		if _, err := os.Stat(filePath); err == nil {
-			return GithubCredentialsFromFile(filePath)
-		}
-	}
-
-	// default XDG_CONFIG_HOME
-	if len(xdgEnvar) == 0 {
-		filePath := path.Join(usr.HomeDir, ".config/snitch/github.ini")
-		if _, err := os.Stat(filePath); err == nil {
-			return GithubCredentialsFromFile(filePath)
-		}
-	}
-
-	filePath := path.Join(usr.HomeDir, ".snitch/github.ini")
-	if _, err := os.Stat(filePath); err == nil {
-		return GithubCredentialsFromFile(filePath)
-	}
-
-	return GithubCredentials{}, fmt.Errorf("GitHub token is missing")
 }
 
 func handleError(err error) {
